backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "An AI study notes application backend."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.34",
    "langchain-openrouter>=0.0.1",
    "openai>=2.1.0",
    "pydantic>=2.11.9",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tensorflow>=2.20.0",
    "textract>=1.6.5",
    "transformers>=4.57.0",
]
